Split out attribution of bonuses
